fix(datasets): lift filter-state isolation to BaseSupersetModelRestApi
Per reviewer feedback on #39685: the FAB `_handle_filters_args` race
that this PR addressed for `DatasetRestApi` exists in every other
subclass too (ChartRestApi, DashboardRestApi, SavedQueryRestApi,
QueryRestApi, DatabaseRestApi, etc.) — they all inherit from FAB's
`ModelRestApi`, which mutates a single shared `self._filters` instance
across requests.

Move the override from `DatasetRestApi` to `BaseSupersetModelRestApi`
so all subclasses inherit the request-scoped behavior. The override is
purely generic — no dataset-specific logic — and no other superset
REST API overrides `_handle_filters_args` (checked), so there's zero
collision risk. `BaseSupersetModelRestApi` is documented as the place
to "implement specific superset generic functionality", so wrapping
FAB behavior we want to differ from upstream is exactly its purpose.

Test stays in `tests/unit_tests/datasets/api_tests.py` using
`DatasetRestApi` as a concrete subclass to exercise the inherited
behavior; docstring updated to note the broader scope.
